+++ packages/block-editor/src/components/block-lock/modal-actions-slot.js
@@ -1,0 +1,10 @@
+/**
+ * WordPress dependencies
+ */
+import { createSlotFill } from '@wordpress/components';
+
+const BlockLockModalActionsSlotFill = createSlotFill(
+	Symbol( 'BlockLockModalActions' )
+);
+
+export default BlockLockModalActionsSlotFill;